# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TITAN API",
    "summary": "Self-hosted AI assistant, planner and tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "The caller's usage for a month",
        "operationId": "my_usage_api_v1_usage_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Calendar month in UTC, YYYY-MM; the current month by default",
              "examples": [
                "2026-09"
              ],
              "title": "Month"
            },
            "description": "Calendar month in UTC, YYYY-MM; the current month by default"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthUsageOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/usage/household": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Every user's usage for a month (owner only)",
        "operationId": "household_usage_api_v1_usage_household_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Calendar month in UTC, YYYY-MM; the current month by default",
              "examples": [
                "2026-09"
              ],
              "title": "Month"
            },
            "description": "Calendar month in UTC, YYYY-MM; the current month by default"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberUsageOut"
                  },
                  "title": "Response Household Usage Api V1 Usage Household Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/usage/budget": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "The caller's budget this month",
        "operationId": "my_budget_api_v1_usage_budget_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/v1/usage/budgets": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Every user's budget this month (owner only)",
        "operationId": "household_budgets_api_v1_usage_budgets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberBudgetOut"
                  },
                  "type": "array",
                  "title": "Response Household Budgets Api V1 Usage Budgets Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/v1/usage/budgets/{user_id}": {
      "put": {
        "tags": [
          "usage"
        ],
        "summary": "Set or remove a user's monthly limit (owner only)",
        "operationId": "set_budget_api_v1_usage_budgets__user_id__put",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberBudgetOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetIn": {
        "properties": {
          "limit_usd": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Usd",
            "description": "Monthly limit in US dollars, rounded to cents; null removes the cap"
          },
          "owner_alerts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerAlerts"
              },
              {
                "type": "null"
              }
            ],
            "description": "Left out: unchanged, or exceeded for a new limit. Needs a limit"
          }
        },
        "type": "object",
        "required": [
          "limit_usd"
        ],
        "title": "BudgetIn"
      },
      "BudgetOut": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month",
            "description": "The current calendar month in UTC"
          },
          "limit_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Usd",
            "description": "The monthly limit; null means no cap"
          },
          "spent_usd": {
            "type": "number",
            "title": "Spent Usd",
            "description": "The month's cost so far, as in the usage totals"
          },
          "state": {
            "$ref": "#/components/schemas/BudgetState",
            "description": "ok below 80 % of the limit, warning from 80 %, exceeded from 100 %"
          },
          "owner_alerts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerAlerts"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which states the owners are notified of: off, exceeded, or all (warning and exceeded); null without a limit"
          }
        },
        "type": "object",
        "required": [
          "month",
          "limit_usd",
          "spent_usd",
          "state",
          "owner_alerts"
        ],
        "title": "BudgetOut"
      },
      "BudgetState": {
        "type": "string",
        "enum": [
          "ok",
          "warning",
          "exceeded"
        ],
        "title": "BudgetState"
      },
...
      "MemberBudgetOut": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month",
            "description": "The current calendar month in UTC"
          },
          "limit_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Usd",
            "description": "The monthly limit; null means no cap"
          },
          "spent_usd": {
            "type": "number",
            "title": "Spent Usd",
            "description": "The month's cost so far, as in the usage totals"
          },
          "state": {
            "$ref": "#/components/schemas/BudgetState",
            "description": "ok below 80 % of the limit, warning from 80 %, exceeded from 100 %"
          },
          "owner_alerts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerAlerts"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which states the owners are notified of: off, exceeded, or all (warning and exceeded); null without a limit"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "month",
          "limit_usd",
          "spent_usd",
          "state",
          "owner_alerts",
          "user_id",
          "username"
        ],
        "title": "MemberBudgetOut"
      },
      "MemberUsageOut": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "total": {
            "$ref": "#/components/schemas/TotalsOut"
          },
          "by_model": {
            "items": {
              "$ref": "#/components/schemas/ModelTotalsOut"
            },
            "type": "array",
            "title": "By Model"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "month",
          "total",
          "by_model",
          "user_id",
          "username"
        ],
        "title": "MemberUsageOut"
      },
...
      "ModelTotalsOut": {
        "properties": {
          "sessions": {
            "type": "integer",
            "title": "Sessions",
            "description": "Agent sessions that reported usage"
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens"
          },
          "cache_read_tokens": {
            "type": "integer",
            "title": "Cache Read Tokens"
          },
          "cache_creation_tokens": {
            "type": "integer",
            "title": "Cache Creation Tokens"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "description": "As Claude Code reports it; in oauth mode, what the tokens would cost on the API"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "sessions",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_creation_tokens",
          "cost_usd",
          "model"
        ],
        "title": "ModelTotalsOut"
      },
      "MonthUsageOut": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "total": {
            "$ref": "#/components/schemas/TotalsOut"
          },
          "by_model": {
            "items": {
              "$ref": "#/components/schemas/ModelTotalsOut"
            },
            "type": "array",
            "title": "By Model"
          }
        },
        "type": "object",
        "required": [
          "month",
          "total",
          "by_model"
        ],
        "title": "MonthUsageOut"
      },
...
      "OwnerAlerts": {
        "type": "string",
        "enum": [
          "off",
          "exceeded",
          "all"
        ],
        "title": "OwnerAlerts",
        "description": "Which of a user's budget states the owners are notified of."
      },
...
      "TotalsOut": {
        "properties": {
          "sessions": {
            "type": "integer",
            "title": "Sessions",
            "description": "Agent sessions that reported usage"
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens"
          },
          "cache_read_tokens": {
            "type": "integer",
            "title": "Cache Read Tokens"
          },
          "cache_creation_tokens": {
            "type": "integer",
            "title": "Cache Creation Tokens"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "description": "As Claude Code reports it; in oauth mode, what the tokens would cost on the API"
          }
        },
        "type": "object",
        "required": [
          "sessions",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_creation_tokens",
          "cost_usd"
        ],
        "title": "TotalsOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Device token from POST /devices/pair",
        "scheme": "bearer"
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "description": "Set by POST /session; browsers only",
        "in": "cookie",
        "name": "__Host-TSID"
      }
    }
  }
}